pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-job-hunter"
version = "1.1.0"
description = "Automated job hunting pipeline with AI support — collect, AI-filter, and notify."
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.11"
authors = [{ name = "Nicolas Alcaraz" }]
keywords = ["jobs", "automation", "ai", "pipeline", "minimax"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "python-jobspy>=1.1.0",
    "anthropic>=0.40.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
job-hunter = "job_hunter.main:main"

[project.urls]
Repository = "https://github.com/dablon/ai-job-hunter"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
